Started change to new framework

+++ src/Spirebyte.Services.Repositories.Core/Repositories/IPullRequestRepository.cs
@@ -5,11 +5,11 @@ namespace Spirebyte.Services.Repositories.Core.Repositories;
 
 public interface IPullRequestRepository
 {
-    Task<PullRequest> GetAsync(string repositoryId, int pullRequestId);
+    Task<PullRequest> GetAsync(string repositoryId, long pullRequestId);
     Task<long> GetPullRequestCountOfRepositoryAsync(string repositoryId);
-    Task<bool> ExistsAsync(string repositoryId, int pullRequestId);
+    Task<bool> ExistsAsync(string repositoryId, long pullRequestId);
     Task AddAsync(string repositoryId, PullRequest pullRequest);
     Task UpdateAsync(string repositoryId, PullRequest pullRequest);
 
-    Task DeleteAsync(string repositoryId, int pullRequestId);
+    Task DeleteAsync(string repositoryId, long pullRequestId);
 }
